+++ models/template.py
@@ -1,0 +1,6 @@
+import tensorflow as tf
+from tensorflow import keras
+
+def model(shape=(256, 512, 3)):
+    """Template model to create keras models"""
+    pass
